Made a folder for Obesity dataSet.
Added the linear regression Model. First version does only include one
loss function. Works for simple example from the scikit-learn page for
linear regression. Does encounter a division by zero if the weights lead
to the exact equation for y, but that should not be an issue with real
data sets.
Output weights are ordered the same as for the scikit learn linear
regression, but the third weight in my output is the 'reg.intercept_'
parameter in the scikit learn implementation.