Added the option to have alpha change over time and to set an alpha for each weight. Also added Mean Square Error (MSE) as well as Mean Absolute Error (MAE) for loss functions.
Included a version off the linear regression,
that solves the minimization problem of the square
error in form of a matrix equation (uses scipy.linalg.solve(). To show,
that that is indeed how the scikit learn linear regression solves the
problem. This does not need a learning rate or initial weigths and
therefore is a lot more stable and faster...

Also the code is extremely slow compared to the matrix version. Could
probably be optimized a lot because I used a lot of "unneccessary" function
calls for the sake of readability.

The linearRegression() calssifier is EXTREMELY sensitive to initial
weights and choice of alpha and only converges when alpha is chosen
well.